requirements.txt
ccxt==4.3.69
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
tenacity==8.5.0
python-dateutil==2.9.0.post0
